+++ tstest/integration/tailscaled_deps_test_windows.go
@@ -27,7 +27,8 @@ import (
 	_ "tailscale.com/net/netns"
 	_ "tailscale.com/net/portmapper"
 	_ "tailscale.com/net/proxymux"
-	_ "tailscale.com/net/socks5/tssocks"
+	_ "tailscale.com/net/socks5"
+	_ "tailscale.com/net/tsdial"
 	_ "tailscale.com/net/tshttpproxy"
 	_ "tailscale.com/net/tstun"
 	_ "tailscale.com/paths"
@@ -36,6 +37,7 @@ import (
 	_ "tailscale.com/types/flagtype"
 	_ "tailscale.com/types/key"
 	_ "tailscale.com/types/logger"
+	_ "tailscale.com/types/netmap"
 	_ "tailscale.com/util/clientmetric"
 	_ "tailscale.com/util/multierr"
 	_ "tailscale.com/util/osshare"
