+++ wgengine/netstack/netstack.go
@@ -33,10 +33,10 @@ import (
 	"inet.af/netstack/waiter"
 	"tailscale.com/net/packet"
 	"tailscale.com/net/tsaddr"
+	"tailscale.com/net/tsdial"
 	"tailscale.com/net/tstun"
 	"tailscale.com/types/logger"
 	"tailscale.com/types/netmap"
-	"tailscale.com/util/dnsname"
 	"tailscale.com/wgengine"
 	"tailscale.com/wgengine/filter"
 	"tailscale.com/wgengine/magicsock"
@@ -79,7 +79,7 @@ type Impl struct {
 	atomicIsLocalIPFunc atomic.Value // of func(netaddr.IP) bool
 
 	mu  sync.Mutex
-	dns DNSMap
+	dns tsdial.DNSMap
 	// connsOpenBySubnetIP keeps track of number of connections open
 	// for each subnet IP temporarily registered on netstack for active
 	// TCP connections, so they can be unregistered when connections are
@@ -179,59 +179,10 @@ func (ns *Impl) Start() error {
 	return nil
 }
 
-// DNSMap maps MagicDNS names (both base + FQDN) to their first IP.
-// It should not be mutated once created.
-type DNSMap map[string]netaddr.IP
-
-func DNSMapFromNetworkMap(nm *netmap.NetworkMap) DNSMap {
-	ret := make(DNSMap)
-	suffix := nm.MagicDNSSuffix()
-	have4 := false
-	if nm.Name != "" && len(nm.Addresses) > 0 {
-		ip := nm.Addresses[0].IP()
-		ret[strings.TrimRight(nm.Name, ".")] = ip
-		if dnsname.HasSuffix(nm.Name, suffix) {
-			ret[dnsname.TrimSuffix(nm.Name, suffix)] = ip
-		}
-		for _, a := range nm.Addresses {
-			if a.IP().Is4() {
-				have4 = true
-			}
-		}
-	}
-	for _, p := range nm.Peers {
-		if p.Name == "" {
-			continue
-		}
-		for _, a := range p.Addresses {
-			ip := a.IP()
-			if ip.Is4() && !have4 {
-				continue
-			}
-			ret[strings.TrimRight(p.Name, ".")] = ip
-			if dnsname.HasSuffix(p.Name, suffix) {
-				ret[dnsname.TrimSuffix(p.Name, suffix)] = ip
-			}
-			break
-		}
-	}
-	for _, rec := range nm.DNS.ExtraRecords {
-		if rec.Type != "" {
-			continue
-		}
-		ip, err := netaddr.ParseIP(rec.Value)
-		if err != nil {
-			continue
-		}
-		ret[strings.TrimRight(rec.Name, ".")] = ip
-	}
-	return ret
-}
-
 func (ns *Impl) updateDNS(nm *netmap.NetworkMap) {
 	ns.mu.Lock()
 	defer ns.mu.Unlock()
-	ns.dns = DNSMapFromNetworkMap(nm)
+	ns.dns = tsdial.DNSMapFromNetworkMap(nm)
 }
 
 func (ns *Impl) addSubnetAddress(ip netaddr.IP) {
@@ -345,49 +296,6 @@ func (ns *Impl) updateIPs(nm *netmap.NetworkMap) {
 			ns.logf("[v2] netstack: registered IP %s", ipp)
 		}
 	}
-}
-
-// Resolve resolves addr into an IP:port using first the MagicDNS contents
-// of m, else using the system resolver.
-func (m DNSMap) Resolve(ctx context.Context, addr string) (netaddr.IPPort, error) {
-	ipp, pippErr := netaddr.ParseIPPort(addr)
-	if pippErr == nil {
-		return ipp, nil
-	}
-	host, port, err := net.SplitHostPort(addr)
-	if err != nil {
-		// addr is malformed.
-		return netaddr.IPPort{}, err
-	}
-	if net.ParseIP(host) != nil {
-		// The host part of addr was an IP, so the netaddr.ParseIPPort above should've
-		// passed. Must've been a bad port number. Return the original error.
-		return netaddr.IPPort{}, pippErr
-	}
-	port16, err := strconv.ParseUint(port, 10, 16)
-	if err != nil {
-		return netaddr.IPPort{}, fmt.Errorf("invalid port in address %q", addr)
-	}
-
-	// Host is not an IP, so assume it's a DNS name.
-
-	// Try MagicDNS first, else otherwise a real DNS lookup.
-	ip := m[host]
-	if !ip.IsZero() {
-		return netaddr.IPPortFrom(ip, uint16(port16)), nil
-	}
-
-	// No MagicDNS name so try real DNS.
-	var r net.Resolver
-	ips, err := r.LookupIP(ctx, "ip", host)
-	if err != nil {
-		return netaddr.IPPort{}, err
-	}
-	if len(ips) == 0 {
-		return netaddr.IPPort{}, fmt.Errorf("DNS lookup returned no results for %q", host)
-	}
-	ip, _ = netaddr.FromStdIP(ips[0])
-	return netaddr.IPPortFrom(ip, uint16(port16)), nil
 }
 
 func (ns *Impl) DialContextTCP(ctx context.Context, addr string) (*gonet.TCPConn, error) {
